+++ simple-stack-flow-masterdetail/src/main/java/com/example/stackmasterdetail/MainActivity.java
@@ -24,7 +24,7 @@ import android.view.MenuItem;
 
 import com.example.stackmasterdetail.pathview.HandlesBack;
 import com.example.stackmasterdetail.util.BackstackService;
-import com.example.stackmasterdetail.util.MasterDetailBackstackDelegate;
+import com.example.stackmasterdetail.util.MasterDetailStateClearStrategy;
 import com.zhuinden.simplestack.BackstackDelegate;
 import com.zhuinden.simplestack.HistoryBuilder;
 import com.zhuinden.simplestack.StateChange;
@@ -64,7 +64,8 @@ public class MainActivity
         super.onCreate(savedInstanceState);
         final ActionBar actionBar = getSupportActionBar();
         actionBar.setDisplayShowHomeEnabled(false);
-        backstackDelegate = new MasterDetailBackstackDelegate(null);
+        backstackDelegate = new BackstackDelegate(null);
+        backstackDelegate.setStateClearStrategy(new MasterDetailStateClearStrategy());
         backstackDelegate.onCreate(savedInstanceState,
                 getLastCustomNonConfigurationInstance(),
                 HistoryBuilder.single(Paths.ConversationList.create()));